Move ObjDict.
